refactor integration test helpers into conda.testing
