+++ tests/conda_env/test_create.py
@@ -18,7 +18,7 @@ from conda.models.enums import PackageType
 from conda.models.match_spec import MatchSpec
 from . import support_file
 from .utils import make_temp_envs_dir, Commands, run_command
-from ..test_create import run_command as run_conda_command, \
+from conda.testing.integration import run_command as run_conda_command, \
     Commands as CondaCommands
 
 PYTHON_BINARY = 'python.exe' if on_win else 'bin/python'
