+++ tests/test_export.py
@@ -5,7 +5,7 @@ from unittest import TestCase
 from conda.gateways.disk.delete import rm_rf
 import pytest
 
-from .test_create import Commands, PYTHON_BINARY, make_temp_env, make_temp_prefix, \
+from conda.testing.integration import Commands, PYTHON_BINARY, make_temp_env, make_temp_prefix, \
     package_is_installed, run_command
