+++ tests/test_history.py
@@ -5,7 +5,7 @@ import unittest
 from conda.testing.cases import BaseTestCase
 from conda.testing.decorators import skip_if_no_mock
 from conda.testing.helpers import mock
-from .test_create import make_temp_prefix
+from conda.testing.integration import make_temp_prefix
 
 from conda.history import History
 from conda.common.compat import text_type
