+++ tests/test_priority.py
@@ -7,7 +7,7 @@ import pytest
 from conda.base.context import context, conda_tests_ctxt_mgmt_def_pol
 from conda.common.compat import on_win
 from conda.common.io import env_var
-from .test_create import Commands, package_is_installed, get_conda_list_tuple, \
+from conda.testing.integration import Commands, package_is_installed, get_conda_list_tuple, \
     make_temp_env, run_command
